+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN9Z7DFAFWMTHJ55HV9Y97AY"
+        "01KNGMV8YXCGYKNA3C9GKY77GY"
     )
     r"""The unique identifier of the data source"""