+++ packages/orq-rc/RELEASES.md
@@ -6859,3 +6859,13 @@ Based on:
 - [python v4.7.0-rc.46] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.46] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.46 - packages/orq-rc
+
+## 2026-04-06 05:42:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.47] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.47 - packages/orq-rc